Extract learn more link into reusable component
